timerfd: support TFD_NONBLOCK flag
Currently, the timerfd module does not allow specifying different flags
for the created timers. Add support for this via a new method,
`TimerFd::new_with_flags`.

This will allow firecracker to use the implementation from this crate,
instead of the dedicated `timerfd` crate (which pulls in the rustix
crate, of which we have like 3 different versions in our dependency tree
right now).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>